requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.1
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
copier==8.3.0
dateparser==1.2.0
decorator==5.1.1
document==1.0
docx==0.2.4
dunamai==1.19.0
exceptiongroup==1.1.3
Flask==3.0.3
flet==0.10.3
flet-core==0.10.3
flet-runtime==0.10.3
frozenlist==1.4.1
funcy==2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jinja2-ansible-filters==1.3.2
lxml==5.2.1
macholib==1.16.3
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pathspec==0.11.2
pillow==10.3.0
plumbum==1.8.2
prompt-toolkit==3.0.36
pycryptodome==3.20.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pypng==0.20220715.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
redis==5.0.4
regex==2023.12.25
repath==0.9.0
reportlab==4.2.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
watchdog==3.0.0
wcwidth==0.2.8
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==3.0.2
yarl==1.9.4
